Refactor RuleList
