+++ src/discounts/components/DiscountRules/componenets/RulesList/RulesList.tsx
@@ -1,12 +1,6 @@
 import { Rule } from "@dashboard/discounts/models";
 import { CommonError } from "@dashboard/utils/errors/common";
-import {
-  Box,
-  Button,
-  EditIcon,
-  Text,
-  TrashBinIcon,
-} from "@saleor/macaw-ui-next";
+import { Box, Text } from "@saleor/macaw-ui-next";
 import React from "react";
 import { useIntl } from "react-intl";
 
@@ -14,6 +8,8 @@ import { useDiscountRulesContext } from "../../context";
 import { messages } from "../../messages";
 import { getCurencySymbol } from "../../utils";
 import { Placeholder } from "../Placeholder";
+import { RuleActions } from "./components/RuleActions";
+import { RuleLabel } from "./components/RuleLabel";
 import { RuleListContainer } from "./components/RuleListContainer";
 import { RuleListLoading } from "./components/RuleListLoading";
 import { RuleSummary } from "./components/RuleSummary";
@@ -21,6 +17,7 @@ import { RuleWrapper } from "./components/RuleWrapper";
 
 interface RulesListProps<ErrorCode> {
   rules: Rule[];
+  disabled?: boolean;
   errors: Array<CommonError<ErrorCode> & { index?: number }>;
   loading?: boolean;
   onRuleDelete: (index: number) => void;
@@ -35,7 +32,7 @@ export const RulesList = <ErrorCode,>({
   loading,
 }: RulesListProps<ErrorCode>) => {
   const intl = useIntl();
-  const { channels, disabled } = useDiscountRulesContext();
+  const { channels } = useDiscountRulesContext();
 
   if (loading) {
     return <RuleListLoading />;
@@ -44,13 +41,6 @@ export const RulesList = <ErrorCode,>({
   if (rules.length === 0) {
     return <Placeholder />;
   }
-
-  const getRuleName = (name: string | undefined) => {
-    if (name) {
-      return `: ${name}`;
-    }
-    return "";
-  };
 
   return (
     <RuleListContainer>
@@ -65,33 +55,12 @@ export const RulesList = <ErrorCode,>({
                 justifyContent="space-between"
                 alignItems="center"
               >
-                {intl.formatMessage(messages.catalogRule) +
-                  getRuleName(rule.name)}
+                <RuleLabel ruleName={rule.name} />
 
-                <Box display="flex">
-                  <Button
-                    size="small"
-                    variant="tertiary"
-                    onClick={() => onRuleEdit(index)}
-                    cursor={disabled ? "not-allowed" : "pointer"}
-                    disabled={disabled}
-                    data-test-id="rule-edit-button"
-                  >
-                    <EditIcon />
-                  </Button>
-                  <Button
-                    size="small"
-                    disabled={disabled}
-                    variant="tertiary"
-                    data-test-id="rule-delete-button"
-                    onClick={e => {
-                      e.stopPropagation();
-                      onRuleDelete(index);
-                    }}
-                  >
-                    <TrashBinIcon />
-                  </Button>
-                </Box>
+                <RuleActions
+                  onDelete={() => onRuleDelete(index)}
+                  onEdit={() => onRuleEdit(index)}
+                />
               </Box>
               <RuleSummary
                 rule={rule}
