+++ src/discounts/components/DiscountRules/componenets/RulesList/components/RuleLabel/index.ts
@@ -1,0 +1,1 @@
+export * from "./RuleLabel";
